Add Oracle integration tests and numeric ordering for data fields
Ordering by a data path passed no type to Oracle, so numbers were ordered
lexically and 10 came before 2. The strategy now returns a numeric order
key plus a text fallback, and the whole DB search runs against a real
Oracle XE container: numeric ordering and range filters, equality, in,
contains, full text and column filters.

+++ src/test/java/com/icthh/xm/ms/entity/service/search/db/dialect/OracleJsonValueStrategyUnitTest.java
@@ -25,8 +25,8 @@ public class OracleJsonValueStrategyUnitTest extends AbstractJupiterUnitTest {
         assertThat(OracleJsonValueStrategy.returningType(10L)).isEqualTo(Long.class);
         assertThat(OracleJsonValueStrategy.returningType(10)).isEqualTo(Long.class);
         assertThat(OracleJsonValueStrategy.returningType(BigInteger.TEN)).isEqualTo(Long.class);
-        assertThat(OracleJsonValueStrategy.returningType(10.5d)).isEqualTo(Double.class);
-        assertThat(OracleJsonValueStrategy.returningType(10.5f)).isEqualTo(Double.class);
+        assertThat(OracleJsonValueStrategy.returningType(10.5d)).isEqualTo(BigDecimal.class);
+        assertThat(OracleJsonValueStrategy.returningType(10.5f)).isEqualTo(BigDecimal.class);
         assertThat(OracleJsonValueStrategy.returningType(BigDecimal.ONE)).isEqualTo(BigDecimal.class);
         assertThat(OracleJsonValueStrategy.returningType(true)).isEqualTo(Boolean.class);
         assertThat(OracleJsonValueStrategy.returningType("Kyiv")).isEqualTo(String.class);
